+++ android/src/main/java/com/lovegaoshi/kotlinaudio/models/PlayWhenReadyChangeData.kt
@@ -1,3 +1,7 @@
 package com.lovegaoshi.kotlinaudio.models
 
-data class PlayWhenReadyChangeData(val playWhenReady: Boolean, val pausedBecauseReachedEnd: Boolean)
+data class PlayWhenReadyChangeData(
+    val playWhenReady: Boolean,
+    val pausedBecauseReachedEnd: Boolean,
+    val pausedBecauseBecameNoisy: Boolean
+)